requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
decouple==0.0.7
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
jsonify==0.5
numpy==1.23.4
openai==0.24.0
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.5.1
pandas-stubs==1.5.1.221024
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
types-pytz==2022.5.0.0
typing_extensions==4.4.0
urllib3==1.26.12
